Tried USART2 via PuTTY. Debuged with several letters and digits. USART2 receive working

+++ Inc/uart_driver.h
@@ -10,6 +10,6 @@ void UART_RXIE_enable(USARTx_typeDef* USARTx);
 void UART2_init(void);
 void UART_set_baud_rate(USARTx_typeDef* USARTx, uint32_t baudRate, uint32_t periphClock);
 uint8_t UART_check_RXNE_flag(USARTx_typeDef* USARTx);
-uint8_t UART_data_return(USARTx_typeDef* USARTx);
+uint32_t UART_data_return(USARTx_typeDef* USARTx);
 
 #endif /* UART_DRIVER_H_ */
